Changed ints to doubles to allow decimal results

+++ Calculator/StackImplementation.swift
@@ -6,15 +6,15 @@
 import Foundation
 
 class Stack {
-    var stackArray = [Int]()
+    var stackArray = [Double]()
     //create push function
     //create pop function
     
-    func push(intToPush: Int){
+    func push(intToPush: Double){
         self.stackArray.append(intToPush)
     }
     
-    func pop() -> Int? {
+    func pop() -> Double? {
         if self.stackArray.last != nil {
             var stringToReturn = self.stackArray.last
             self.stackArray.removeLast()
